+++ examples/gps.rs
@@ -10,7 +10,7 @@ fn main() {
 
     if let Ok(meta) = Metadata::new_from_path("examples/example.jpg") {
         if let Some(location) = meta.get_gps_info() {
-            println!("Location: {:?}", location);
+            println!("Location: {location:?}");
         }
     }
 }