+++ examples/timestamp.rs
@@ -10,7 +10,7 @@ fn main() {
 
     if let Ok(meta) = Metadata::new_from_path("examples/example.jpg") {
         if let Ok(time) = meta.get_tag_string("Exif.Image.DateTime") {
-            println!("Time: {:?}", time);
+            println!("Time: {time:?}");
         }
         if meta
             .set_tag_string("Exif.Image.DateTime", "2008:11:01 21:15:07")