+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [NEXT] - Unreleased
+  * Require Rust 1.56 as the minimum supported language version.
+  * Adopt 2021 edition of the language.
+  * Dependency upgrades.
+
 ## [v0.9.1] - 2020-07-12
   * Fixed build failure on arm32 due to invalid assumptions about int size.
   * Fixed null pointer crash when using `get_tag_raw()`.