+++ CHANGELOG.md
@@ -4,7 +4,7 @@ SPDX-License-Identifier: CC0-1.0
 -->
 
 ## [NEXT] - Unreleased
-  * Require Rust 1.56 as the minimum supported language version.
+  * Require Rust 1.63 as the minimum supported language version.
   * Adopt 2021 edition of the language.
   * Dependency upgrades.